commit by aliyun ECS

+++ day30-31/process_pool_6.0/server/process_child.h
@@ -4,7 +4,7 @@
 #include <func.h>
 
 struct trans_t {
-    off_t buflen;
+    size_t buflen;
     char buf[1024];
 };
